chore(iast): restructure functions
